refactor: DetailProductScreen 내 하위 컴포넌트 분리 및 패키지 구조 개선
